EWOK-30 Added dynamic recipe tab
Added the file for a dynamic tab that can take a recipe id as a parameter.
The tab takes the id (from the parameter) and uses it to fetch the full recipe information from the recipeContext.
Includes addition of getRecipebyID method to recipeContext (which calls the same method int he service file)