SQUASH 4: EWOK-95 Added return type for comparer
Added exported type that will be returned by the ingredient-wise comparer.

EWOK-95 Added ingredient-wise comparer method

Added exported method that checks the ingredients in a recipe against the ingredients in the given pantry and returns information about whether each ingredient in the recipe is present in the pantry and in sufficient amount.

EWOK-95 Added unit tests for comparer

EWOK-95 Altered return structure from PR request

Altered the structure of the return from the compareRecipeToPantryDetails method to fit the structure rquested in the pull request.