SQUASH 3: EWOK-140 Added filter recipes by pantry switch
Updated the search bar to include a switch to search recipes using the user's pantry.

EWOK-140 Added switch to advanced filter & fixed live update

Added an identical and linked switch to the advanced filter modal.
Added useEffect to hide the switch when the pantry becomes empty.

EWOK-140 Removed "My Pantry" filter chip

Removed the filter chip for the pantry because the switch is visible in the same place, so the chip is redundant.